modification du type de l'id en BigInteger

+++ server/src/main/java/fr/uga/l3miage/tp1/exo2/models/NetflixAccountEntity.java
@@ -1,12 +1,14 @@
 package fr.uga.l3miage.tp1.exo2.models;
 
 import javax.persistence.*;
+import java.math.BigInteger;
+
 @Entity
 @Table(name = "netflix_account")
 public class NetflixAccountEntity {
     @Id
     @Column(name = "id")
-    private long id;
+    private BigInteger id;
     @Column(name = "nb_devices")
     private int nbDevices;
     @Column(name = "typeAccount")
